% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost.R
\docType{package}
\name{adaboost}
\alias{adaboost}
\alias{adaboost-package}
\title{Adaboost is a package for LiDAR data visualization and statistical analysis}
\description{
Adaboost is a package for Light Detection and Ranging(LiDAR) data visualization
and statistical analysis. This package provides a collection of tools which users
can follow through to gain understanding of LiDAR semantics. Adaboost package
incorporates all preliminary data extraction and visulization steps for LiDAR files
of type (.las or .laz or .LAS or .LAZ), provides functions to use a columnar storage
like MonetDBLite to load large data files into database tables, presents a list
of tree scores associated to the file that can be used as input for a Machine
Learning Model related to feature extraction or vegetation segmentation.
}
\author{
Minakshi Mukherjee

Maintainer: Minakshi Mukherjee<adaboost@stanford.edu>
}
\keyword{package}
