% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasSlotExtract.R
\name{lasSlotExtract}
\alias{lasSlotExtract}
\title{Extracts the contents of LAS(LASF) class object and creates tibbles from S4 slots data and header.}
\usage{
lasSlotExtract(lasObject, fileName)
}
\arguments{
\item{lasObject}{A las object of class:LAS (LASF) to be analyzed}

\item{fileName}{A Fie in  (ASPRS) standard .las or .laz or .LAS or .LAZ format}
}
\value{
A list with 9 components: tibbles, colnames, las metadata
}
\description{
This function extracts attributes from point cloud S4 slots Data and Header.
Some of the important attributes of Slot Data are: X, Y, Z values(these determine
point cloud coordinates), Standard GPS time, Intensity to store magnitude of laser
pulse return, Return Number for a given output laser pulse, Number of Returns for a
given laser pulse, Classification representing class attribute of the point
(Ground, Vegetation, Building etc.)
}
\details{
It provides a lookup table to map standard Classification Key(0-9) to its meaning.
Attributes from S4 slot data is stored in a tibble with well defined column names
to facilitatefor any future analysis. Any LiDAR file can potentially contains billions
of point clouds, so S4 slot Data will have billions of attribute entries that represent
all point clouds.

Point cloud S4 slot Header provides some of the important metadata:
File Signature LASF, File Source ID (e.g. original flight line number),
File Creation Year, Number of Point Clouds in the dataset, Min and Max
values of  of X,Y,Z coordinates. There will be only one entry in the slot
Header per file.

This function provides answers to the following commonly asked questions for LiDAR data:
How many points are there in the Point Cloud Data Set?
What are the distinct return Nos?
What are the distinct No of Returns?
What type of Classification does his dataset contain?
What are the top 5 Intensity values?
}
