% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasVisualize.R
\name{lasVisualize}
\alias{lasVisualize}
\title{Creates plots for Digital Elevation Model(DEM), Canopy Height Model(CHM) and TreeTops on CHM}
\usage{
lasVisualize(lasObject, option)
}
\arguments{
\item{lasObject}{A las object of class:LAS (LASF) to be visualized}

\item{option}{A string that indicates the type of the plot}
}
\value{
A plot based on the options provided
}
\description{
Digital Elevation Model(DEM) refers to the bare earth model with respet to vertical datum.
Vertical datum refers a surface of zero elevation to refer heights in LiDAR file.
Adaboost package has function readFile that reads and processes a LiDAR file, it returns a
LAS object that is already height normalized with respect to vertical datum.
Raw LiDAR points are not spaced on a regular grid as they are collected by zigzagging laser
pulses. Hence, point cloud is converted to grid or raster by resampling x,y coordinates of
LiDAR data.Then inverse distance weighted average(IDW) is computed and interpolated over
no pixels(no data).
}
\details{
DEM is a raster or grid of elevation values that can be rendered to show the landscape.
Goal is to Achieve smallest pixel size while minimizing no of pixels with no shots in them.
Smaller pixels are desirable as they capture all extraordinary details in the change of
elevation or the shape of the surface.

This function filters out non-ground points along with trees to build DEM using one of the
available algorithms from raster grid. Experiment shows that optimum result is obtained by
using search radius between .5 meter(m) to 1m.

Canopy Height Model(CHM) regers to the distance between the ground and the top of the trees.
It represents the top of earth's surface which includes the trees that sit on the Earth, hence
CHM is also called Digital Surface Model(DSM).
It finds the height of the highest point and then replaces each point with 8 points around the
original to allow for virtual ‘emulation’ of the fact that each lidar point is more realistically
a disc.

This function uses one of the existing algorithms grid_canopy based on basic triangulation and
rasterization. Points-to-raster(p2r) algorithm with a resolution of 0.5 meters is used.
It replacing each point by a 20-cm radius circle of 8 points.

The function uses one of the available tree detection algorithm using rastor-based method that
shows clear plots of tree tops on CHM.
}
\examples{
#=======================================================================#
# Visualize CHM or DEM or TreeTop on CHM
#=======================================================================#
dFile <- system.file("extdata", "project.las", package = "adaboost")
lasOut <- readFile(dFile)
lasVisualize(lasOut, "plotChm")
lasVisualize(lasOut, "plotDem")
lasVisualize(lasOut, "plotTreeTop")

}
