% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMonetDB.R
\name{loadMonetDB}
\alias{loadMonetDB}
\title{Reads a tibble into a csv file and creates a table into MonetDBLite}
\usage{
loadMonetDB(tibble, colName, tabName)
}
\arguments{
\item{tibble}{Tibble to be written into csv file}

\item{colName}{Column names for the table to load data}

\item{tabName}{Name of the table to load the data}
}
\value{
Prints table name and output queries from MonetDB after loading the data
}
\description{
MonetDBLite is a columnar storage architecture SQL database that runs inside
R. This function reads a tibble and stores it into a csv file.
In order to create a temporary server for columnar storage MonetDBLite, it
create a DBI connection, then it reads the csv file and creates the table
with specified column name.
}
\details{
It also queries MonetDBLite database and provides results of the following
commonly asked queries:
How many total records are in the table?
Print output of select columns from the table.

This function closes the MonetDBLite connection after fecthing the data.
}
\examples{
#=======================================================================#
# Load into MonetDBLite and run queries
#=======================================================================#
MonetDBLite::monetdblite_shutdown()
dFile <- system.file("extdata", "lasc04765150.las", package = "adaboost")
lasOut <- readFile(dFile)
dataTibble <- lasSlotExtract(lasOut, dFile)$dataTibble
dataColName <- lasSlotExtract(lasOut, dFile)$dbColnamesData
loadMonetDB(dataTibble, dataColName, "slotdata" )
MonetDBLite::monetdblite_shutdown()

}
