% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTiff.R
\name{processTiff}
\alias{processTiff}
\title{Reads and plots raster files of type .tif(GeoTIFF image format)}
\usage{
processTiff(fileRas)
}
\arguments{
\item{fileRas}{GeoTIFF image format(.tif)}
}
\value{
A list with components: raster DataFrame and raster object
}
\description{
The function reads raster or gidded data from a .tif file which is a GeoTIFF image
format containing additional georeferencing information for raster metadata.
This is redenred as pixels, where each pixel represents an area on Earth's surface.
}
\details{
The function reads a Geotiff file and returns a SpatialGridDataFrame object after
further processing. A SpatialGridDataFrame is a SpatialGrid object extended with a
data slot occupied by a data.frame object filled with a single band of data
representing elevation in metres. RasterLayer object can be coerced to
SpatialGridDataFrame.

It plots a Digital Surface Model(DSM) from the raster object which is a rectangular
grid showing specific features on the surface. The raster data frame and raster object
returned by the function can be fed into any machine learning algorithm for analysis of
feature extraction and classification.
}
