% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeDensityPlot.R
\name{treeDensityPlot}
\alias{treeDensityPlot}
\title{3D Scatter plot of trees within LAS rectangle}
\usage{
treeDensityPlot(treeX,treeY,treeZ)
}
\arguments{
\item{treeX}{R numeric vector}

\item{treeY}{R numeric vector}

\item{treeZ}{R numeric vector}
}
\value{
Plot of tree density
}
\description{
This function reads the tree coordinates(X, Y, height) and creates a 3D scatter plot
of the trees. This plot helps to visualize, whether it is any of the following
categories:
(very dense forest of tall trees, dense forest of tall trees, tall sparse trees,
very dense bush, small sparse trees).
}
\examples{
#=======================================================================#
# 3D scatter plot of trees from a LAS file
#=======================================================================#
dFile <- system.file("extdata", "project.las", package = "adaboost")
lasOut <- readFile(dFile)
scoreOutput <- treeDensityIndex(lasOut)
treeCoordinates <- scoreOutput["treeCoordinates"]
treeX <- treeCoordinates$treeCoordinates$x
treeY <- treeCoordinates$treeCoordinates$y
treeZ <- treeCoordinates$treeCoordinates$height
treeDensityPlot(treeX,treeY,treeZ)

}
